requirements.txt
langchain
langchain_core
langchain_community
langchain_groq
langchain_google_genai
langchain_ollama
python-dotenv
psycopg2-binary
ipython
sentence-transformers
numpy
pydantic
faiss-cpu
unidiff

# Standard library modules (do not need to be in requirements.txt):
# os, sys, time, json, uuid, re, typing, warnings, pathlib, random